Enable eslint (#96)
jshint was misconfigured or just terrible at detecting issues in this project. Swap it with eslint instead and setup a configuration based on the video.js standard. Fix a number of easy linting issues and then tweak the configuration to warn for most of the others. This is definitely worth another pass but puts us in a much better position than we were with basically no linting at all.