CEA608: support for cc2-cc4, special/extended characters, some formatting
This commit adds several feature enhancements to CEA608 processing.

1. All four CC tracks are now available

There are now four `Cea608Stream` objects attached to the caption
stream. Each one receives its corresponding field and data channel and
then sets up constants used for comparing control codes.

Only two `Cea608Stream`s are considered "active" at any given time -- the
ones corresponding to the currently active data channel. When a control
code is received (which includes the data channel for which it applies),
the streams included in that channel are activated. Since it is unclear
which channel is active when starting to play a source, all caption data
is dropped until a control code is received.

Since a caption data packet includes a `field` key, being 0 or 1, the
packet data is dispatched to either the 0th or 1st active stream. In
this way, caption data is only ever `push`ed to the appropriate stream.

When flushing out the caption cues, there is now a `stream` property of
the cue which identifies its originating stream. Additionally, the
transmuxer now adds a `captionStreams` property to its events which
indicates which streams have captions (in the current event).

2. Special and Extended Characters

Before, the `Cea608Stream` would consider all special and extended
characters (as defined in CEA608) as unsupported control codes. It will
now properly translate those characters to their unicode equivalent.

3. Formatting

Underline and italics will now be rendered. These are tracked internally
by instance variable `this.formatting_`.

Additionally, PAC rows are now properly interpreted, and text is
inserted into the correct row in the buffer. Pop-on captions should now
render line breaks at the correct spot. Instead of removing _all_ blank
lines, only those at the start or end of the caption are removed.

4. Misc Structure

The code is structured such that support for more features can be added
easily. This will make support for more features easier to add in the
future, if it is desired.

Still unsupported:
- colors
- custom whitespacing on line start/end
- row positioning
- paint-on mode