feat: support for partial appends (#264)
This mux.js PR adds support for a partial data pipeline in mux.js. This is useful for features like LHLS (low latency HLS). It also adds support for emitting data, caption, id3Frame, trackinfo, and timingInfo as individual events, instead of waiting until a full segment has been processed. This provides a better mechanism for users to consume data from the partial pipeline, as well as write simpler handlers for processed data.

Although there are a good deal of changes, they exist either in the separate pipeline (the partial directory), or are add-ons to the existing transmuxer pipeline, thus, should all be backwards compatible.

reset and endTimeline were added to the primary transmuxer pipeline to allow for clearing any partially handled data, and to mark the end of a timeline (and clear out similar information).

partialFlush (which emits partialdone) was added for the partial pipeline, so that any completed data can be emitted before the next append is made.