fix: Fix an issue where a rollover in only one audio or video track could cause the other to be dropped. (#285)
We saw an issue where some streams had a discontinuity and DTS rollover in one track (e.g. video), but not in another (e.g. audio). The effect of this was that the audio would be interpreted as being ~23 hours in the past and would be thrown out.

For live streams, this meant that the stream would abruptly halt and ultimately cause the player to time out. For VOD streams - at least, in the test case we have - this means audio would be dropped, but video would continue to play.

The solution here is to use a single, shared TimestampRolloverStream for cases where this is possible.