fix: If there isn't any video OR audio track data, just emit done (#279)
There are scenarios where we parsed timed-metadata but did NOT find any video or audio data. For timing reasons, we can't map the timed-metadata to a proper cue-time until we have received PTS timing data from audio or video tracks. This change causes the transmuxer to hold onto any parsed timed-metadata info instead of emitting it, treating that timed-metadata like it came from the NEXT segment (which hopefully has content) so that we can do a proper mapping to generate the cues' startTime and endTime.