fix: always calculate the earliest allowed dts using baseMediaDecodeTime (#317)
timelineStartInfo.baseMediaDecodeTime is 0 when keepOriginalTimestampOffsets is true. This causes earliestAllowedDts to be much higher than it should be, which means that almost all audio is trimmed when seeking backwards for live stream videos.
Therefor, we should track baseMediaDecodeTime even though we don't intend to use it for most things when keepOriginalTimestampOffsets is true. That way we can use it to get the real timeline start dts.