fix(caption-parser): do not fail when first sei sample is null
Do a null check around lastMatchedSample in caption-parser. While having
the null check makes sense, it's possible the content is malformed or
that we're doing a miscalculation somewhere else in the caption-parser.
So, while this will help with #223, we won't close #223 until we can
spend a bit more time investigating the underlying issue.