Adds mfhd starting sequence option to Transmuxer
Adds an option `firstSequenceNumber` to Transmuxer.

Can be used like this:

new Transmuxer({ firstSequenceNumber: 10 });

This is passed in to VideoSegmentStream and AudioSegmentStream.

When mfhd boxes are constructed, the starting sequence number will be
taken from that option. The default value is zero.

This is useful for generating segmented streams out of order. Out of
oder processing allows for restart, parallelism, or random seeking into
source material.