feat: use bigint for 64 bit ints if needed and available. (#383)
Read values as BigInt when appropriate. Potentially, cases where this happens already results in broken playback, but out of abundance of caution we're marking this as a breaking change.

BREAKING CHANGE: In some cases, mux.js will now be returning a BigInt rather than a regular Number value. This means that consumers of this library will need to add checks for BigInt for optimal operation.